pyproject.toml
[project]
name = "api-app-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.6",
    "pytest>=8.4.1",
    "requests>=2.32.4",
    "sqlmodel>=0.0.24",
    "strawberry-graphql[fastapi]>=0.281.0",
]
